support/supportedversion: factor out the version
Pulling in the API package meant that we pulled in all of
controller-runtime, which meant that importing the version package
silently side-effect registered the `--kubeconfig` flag on the
command-line flag-set.
